+++ colpali_engine/trainer/colmodel_training.py
@@ -1,5 +1,5 @@
-import json
 import os
+import warnings
 from dataclasses import dataclass
 from typing import Callable, Dict, Optional, Tuple
 
@@ -220,20 +220,16 @@ class ColModelTraining:
             max_length=self.config.max_length,
         )
         if self.config.eval_dataset_loader is not None:
+            warnings.warn(
+                "The evaluation slightly differs from `vidore-benchmark` and will be deprecated.",
+                category=DeprecationWarning,
+            )
             for test_name, test_dataset_loading_func in self.config.eval_dataset_loader.items():
                 print(f"Evaluating {test_name}")
                 test_ds = test_dataset_loading_func()
                 metrics = self.eval_dataset(test_ds)
                 all_metrics[test_name] = metrics
                 print(f"Metrics for {test_name}: {metrics}")
-
-                # checkpoint dumps
-                with open(f"{self.config.output_dir}/results.json", "w") as f:
-                    json.dump(all_metrics, f)
-
-        # save results as json
-        with open(f"{self.config.output_dir}/results.json", "w") as f:
-            json.dump(all_metrics, f)
 
     def save(self, config_file):
         # save model